Fixed Array.fill_() for different data types
Array.fill_() was previously broken for data types whose size wasn't 32
bit - that's because the first argument of memset() was a 32 bit
integer, so eg if you tried to fill an 8 bit array it would interleave a
load of zeros.

I fixed this by adding a native function called memtile that accepts a
destination pointer, a source pointer the length of the source data in
bytes, and a number of repeats, then just repeats the source data nreps
times at the destination.

I also had to add a function called float_to_half_bits for converting
floating point numbers to bits you can pack into a uint16 value, as
without this it just converts float values to an integer (which then
gets interpreted as a float16 later on leading to wonky values).