Replace std::initializer_list<> with a custom class
When using a cross-platform standalone C++ compiler we can no longer
easily make use of the <initializer_list> header since it relies on
compiler-specific support and may depend on proprietary headers.

This change works around that by defining a wp::initializer_array<>
class which also accepts a braced-init-list on construction but stores
the elements by value. The length of the array is also a template
parameter (instead of being variadic) which can be implicitly deduced
and helps with type safety.