Fix 64-bit abs() implementation
stdlib.h only officially defines abs(int). The MSVC version of the
header also provides an abs(long long) overload when __cplusplus is
defined, but this shouldn't be relied on and we should use llabs()
instead.

Note that alternatively we could also have crt.h define all the C++
overloads, but this adds a lot of code, it's easy to make mistakes, and
it wouldn't be the C RunTime any more. Also we're already using
differently named functions like log() and logf() in builtin.h which is
providing many of these overloads (but not for everything, e.g. only
sqrt() is used and not sqrtf(), but this is safe).