Silence implicit conversion warning
0xffffffff cannot be represented in 32-bit floating-point, which has a
mantissa of only 23 bits. It gets rounded up by 1, to 2^32. This is fine
for the randf() function, which is defined to "Return a random float
between [0.0, 1.0)".

Using the hexadecimal floating-point representation of 2^32 is not
widely understood, and casting a 64-bit integer to a 32-bit float might
be confusing, so just silence the warning by making the cast explicit.