Work around error on loading empty DLL
warp.tests.test_reload produces a DLL with zero sections for
test_reference_reference.py. That's because the latter only contains a
wp.func definition and the C++ file that gets generated from it only
contains static functions. Thus there is nothing for the DLL to export.
While it gets created successfully by LLD, it fails to load.

When using the MSVC linker we don't run into this issue because it
produces some debug information in the .rdata section by default. This
behavior can be disabled using the /NOCOFFGRPINFO flag, which also
results in producing a DLL which fails to load.

Note this is also affected by the /NOENTRY and /NODEFAULTLIB flags,
which we're using to avoid any dependency on the CRT. Without these the
CRT will add at least an entry point function to the DLL.

This issue is worked around by adding an unused _wp_main() function.