Support standalone CPU kernel compilation for macOS
This change makes passing --standalone=1 to build_lib.py work on macOS
as well, producing and using the libclang.dylib library.

Since the linker on macOS doesn't support --start-group/--end-group we
repeat the LLVM archives to link to avoid unresolved symbols. Once
appears to be enough.

Darwin prefixes every symbol with an underscore.