Fix dynamic library linking on macOS
GCC/Clang on macOS only supports linking of dynamic libraries using the
-lx command line option, which links libx.dylib. There is no equivalent
of the -l:x.so option like on Linux which allows omitting the 'lib'
prefix.

This wasn't a problem before because warp.dylib was only loaded
dynamically through Python, and never linked. Linking became necessary
for isolating a standalone subset of the CRT (see warp/native/crt.h).

Future changes for standalone JIT-compilation will eliminate the need to
link libwarp.dylib, but for now this restores supporting the host C++
compiler for compiling CPU compute kernels on macOS.