Support function overriding
This change was initiated due to running into an issue when reloading
transient modules that contain Warp functions.

Doing so would trigger a “duplicate function overload” error due to
the `wp.context.get_module()` function not being able to detect any
code update made to the corresponding Python module, and hence not
being able to clear the cached data from the `wp.context.Module`
instance.

As a result, the `@wp.func` decorator would try to register
the underlying function into an `wp.context.Module` instance that was
already listing the same function since it wasn't cleared as expected.

This patch indirectly address this issue by supporting overriding
functions that share the same signature, instead of trying to
register them as overloads.