Fix symbol name mangling for macOS
The LLJIT::lookup() method prefixes symbols with an underscore on
macOS, while the absolute symbols we define previously didn't have an
underscore.

Note that using LLJIT::lookupLinkerMangled() instead doesn't work
because the kernel functions defined in the .cpp file receive an
underscore (on macOS) from Clang. Also there might be other subtle
issues if we don't conform to the platform's mangling rules.