Remove time resolution sensitive caching
The caching of DLLs or object files checked whether they were newer or
had the same file timestamp as their .cpp files to avoid recompilation.
This fails when the codegen is very fast and a logically newer .cpp file
ends up with the same file timestamp as a previously produced DLL or
object file due to limited time resolution.

This scenario happens in the test_reload.py's test_redefine() test,
where a new version of a .cpp is rapidly generated and we may end up
loading the binary code for it.

Note that while this could be fixed by not skipping compilation when the
timestamps match, we don't actually need this caching mechanism in
build_dll() because we already perform effective hash-based caching in
Module.load().