Allow redefinition of unrolled loop iterators
Loop iterators in kernels are allowed to be redefined within the loop
body. But this previously failed when two unrolled loops "sandwich" a
dynamic loop. That's because the outer loop will define the iterator as
a constant on each iteration, the dynamic loop checks that no previously
defined constants got redefined, but the inner unrolled loop assigns new
constant values to the same iterator variable. This resulted in an
"Error mutating a constant..." exception.

This change addresses the issue by ensuring the iterator of unrolled
loops is treated as a mutable variable instead of a constant. This is
accomplished with an integer cast operator. It doesn't affect the value
but turns the constant into a variable.