Fix support for redefined variables in loops with break
When loops have side-effects we must copy the last assigned values back
to their respective variables before jumping out of the loop with a
break statement.

This change refactors the duplicate logic for this storing of SSA values
into their original variables into a `materialize_redefinitions()`
method. It is reused between for-loop, while-loop, and break statement
handling.

Note that this violates SSA form and thus doesn't support
differentiation. But side-effects in loops weren't supported for
backward mode before so this just follows suit.