Optimize Clang/LLVM debug builds
To support debugging JIT-compiled C++ code we need a debug build of
LLVM so that GDB or LLDB can put a breakpoint on the
`__jit_debug_register_code()` function. They use that as a mechanism to
observe when newly compiled code has been loaded into the process's
memory and needs to have its debug info loaded.

Unfortunately debug builds of LLVM are excruciatingly slow and also
consume inordinate amounts of memory during the build. But we don't need
extensive debuggability of the LLVM code itself when we merely want to
debug Warp kernels (not beyond ensuring the JIT code registration hook
is observable). Therefore we can build Clang/LLVM with optimizations
enabled, while still generating debug info as well.

This change introduces the `--debug_llvm` build flag for Warp, which
defaults to false to enable the above mentioned optimizations. But when
a full debug build of Clang/LLVM is desired one can use
`--debug_llvm =1` to obtain the old behavior.

Note that this only has an effect on builds that use `--build_llvm=1`.
Without that flag we use the Clang/LLVM libraries fetched from Packman,
which are fully optimized and don't contain any debug info.