Generate a .pdb debug database for --mode=debug
Debugging JIT-compiled CPU kernels on Windows requires a warp-clang.pdb
file with debug information for the `__jit_debug_register_code` and
`__jit_debug_descriptor` symbols that are forward declared in
`clang.cpp`. But the prebuilt Clang/LLVM libraries fetched from packman
are release builds, which require a non-debug runtime and `NDEBUG` to be
defined, which was previously accomplished by building all of
`warp-clang.dll` in release mode.

This change ensures that `clang.cpp` gets built in debug mode when
`--mode=debug` was specified, and a `.pdb` file gets generated at link
time, while keeping the release runtime and `NDEBUG` when using the
prebuilt LLVM libraries.

Similarly `warp.dll` gets built with the release runtime and `NDEBUG`
defined even in debug mode, when linking the CUDA libraries. To
distinguish a debuggable build even though `NDEBUG` is defined, a new
`WP_ENABLE_DEBUG` macro is added, which can be queried from the DLL.