Fix disabling asserts in release mode
`assert()` is enabled unless `NDEBUG` is defined. Previously we never
defined it for the standalone JIT-compiler so asserts were inadvertently
enabled in release builds.

Also make sure to use the run-time config mode instead of the build
mode used during building `warp.dll` and `warp-clang.dll`. Note that
this confusion will be gone once support for host-compiler builds of CPU
kernels is removed.