Set the distribution package platform based on the warp/bin libraries
When creating a distribution package with `python3 -m build --wheel` it
previously assumed the libraries in warp/bin to just be data. That
resulted in packages named `warp_lang-{version}-py3-none-any.whl`, which
works cross-platform as long as libraries for all platforms are
included.

Recently this became impractical due to exceeding the PyPi 100 MB limit
when including the standalone Clang/LLVM compiler for Windows, Linux,
and macOS.

This change makes it easier to create platform-specific packages which
adhere to the compatibility tags convention laid out in
https://packaging.python.org/en/latest/specifications/platform-compatibility-tags
by inspecting the file extension of the libraries in the `warp/bin`
directory, creating a package with the correct platform tag, independent
of the system on which the package is built.