Eliminate the need for WP_CPU and WP_CUDA
The `__CUDA_ARCH__` macro is only defined when compiling functions for
the `__device__` execution space. It corresponds mostly with our
previous use of the `WP_CUDA` macro, except that the latter was also
defined when compiling CUDA code for the `__host__` execution space, and
the `CUDA_CALLABLE` macro is defined as both `__host__ __device__`. The
use of `WP_CUDA` in `atomic_add()` which contains CUDA builtin functions
only worked because it's a template function which gets instantiated
as needed.

Hence `__CUDA_ARCH__` is a safer way to discern whether `CUDA_CALLABLE`
functions are being compiled for the host or device execution space.