Write attributes onto USD only when necessary
Ideally we'd only write attributes directly to Fabric without
ever needing to touch the USD stage but, while GPU interop is still
being developed, we workaround this by writing to USD only when their
topology changes.

Additionally we only write the attributes that have their values
updated.

This is implemented as a node that inspects changes in bundles
and accordingly feed the `WritePrims` node.

The result is a large performance increase in some cases.