Update hashes immediately after generating code
Previously the hash was updated after loading the generated code. This
can lead to scenarios where the load fails and the hash is for an older
version of the same source file. If we then revert to this previously
successfully loaded source it will use the wrong binary and continue to
fail to load.

In theory, writing the `.ptx` and `.hash` files should be one atomic
operation so we can allow concurrent accesses to the cache.