Explicitly specify CUDA execution spaces
NVRTC supports a `--device-as-default-execution-space` compilation flag
which allows omitting `__device__`, but this isn't standard CUDA and is
not supported by either NVCC nor Clang/LLVM.

Only a few methods were missing execution space specifiers. Note that
while `CUDA_CALLABLE` is defined as `__host__ __device__` when
`__CUDACC__` is defined and thus isn't an exact semantic replacement for
the NVRTC flag, the `__host__` part is harmless for standard C++ code
and the use of this macro maintains consistency with other Warp code.