Print both stdout and stderr on command failure
The `CalledProcessError` that gets raised by `subprocess.check_output()`
when the return code is non-zero contains both the `stdout` and `stderr`
streams (note that `output` is an alias for `stdout`). It is useful to
print both to help analyze an error.

Since either of them can be `None` when nothing was captured, guards
against that have been added.