Turn `omni.warp` into a Python namespace package
As a result, no `__init__.py` file should be placed under
the directory `exts/omni.warp/omni/warp`.

This is needed because otherwise there are issues loading
the extensions correctly when launching Kit with the flag
`--/app/extensions/fastImporter/enabled=0`, but eventually
the behaviour for the flag set to the value `1`, which is
the default in Kit, might have been patched to result in
the same issue.