Fix not unrolling loops containing break or continue
The handling of break and continue statements in Warp kernels assumed
a dynamic loop was being generated, so we could jump to the test block
or exit block. Unrolled loops don't have these blocks.

Since leaving the loop unrolled but adding these blocks would result in
requiring materialization or phi statements, it's probably not much of a
win over just not unrolling these loops.