Implement native half-precision float support for CPU kernels
The implementation of `wp::half` uses PTX intrinsics to convert between
binary32 and binary16 floating-point values when compiling a kernel
targeting the GPU. But for the CPU it previously used a pure C++
emulation of the binary level computations performed in these
conversions.

This change instead relies on Clang's support for the `_Float16` native
half-precision floating-point type. It slightly reduces the cost of
parsing the emulation code in the header, as well as the cost of
inline expansion, and uses fast CPU instructions to perform the
conversion, analogous to the PTX ones.

The x86 F16C extension provides instructions for converting between
binary32 and binary16 floating-point formats. The last CPU micro-
architecture to not support this extension was discontinued in 2013.
Thus we can safely assume this extension to be ubiquitous.

Support for `wp.half` is also needed outside of kernels. This is
implemented in `warp.cpp` and since this isn't typically compiled with
Clang it can't use `_Float16` and uses the explicit emulation. While we
could use F16C instructions here as well, the overhead of Python makes
the gains insignificant.