Move the exports.h namespace declaration to the header itself
The `exports.h` header is only ever included by `warp.cpp`, where it was
surrounded by `namespace wp { ... }` and `extern "C" { ... }`. It's much
clearer to just have these in the header itself.

Note that the `extern "C"` strips away the namespace for the DLL export
names. Only internal name resolution and linkage uses the namespace.

`exports.h` still isn't self-contained after this change, but it's a
step in the right direction.