Refactor function overload matching
The old code was using a `match` Boolean which does not indicate whether
a match has been found, but whether a mismatch has not been encountered
yet.

This change improves readability by instead only assigning
`resolved_func` when a match was found, and otherwise continuing to the
next overload. The inner loop was refactored into a local function to
avoid the reliance on `match` and simplify the flow control.

Also, checking the output dimensions was made to no longer throw an
exception for very common mismatches.