Don't add default libraries to the pip wheel
When building wheels with `pyproject.toml` containing the instructions for
discovering packages, `setup.py` gets called for the `egg_info` command
without passing in the `--platform` argument (as before), but the default
library binaries get added to the package, in addition to the selected
platform-specific ones.

This can be avoided by ensuring that `wheel_platform` only gets determined
when `setup.py` is called for the `bdist_wheel` command.