Implement basic lvalue support
Previously only struct members and array elements could be used as the
left-hand-side of assignments, besides wholesale local variables. This
change adds support for things like members of structs within arrays.

The new `Reference` class represents a reference type and wraps the type
of the underlying data being referenced. References are implemented at
the C++ level as pointers (notably not as C++ references since their
address can not be assigned, only initialized).

Obtaining the address of an element or member (i.e. pointer arithmetic)
is not differentiable, so for reverse mode essentially the same code is
emitted as before. This requires knowing for each variable declaration
and access being emitted whether it's a primal variable or adjoint.