Fix type() operator for references
Evaluating an expression like `type(array[0])` produces a `Reference`
type for the argument. We want to obtain the underlying value type
instead. References don't exist at the Warp kernel language level.

Note this also means that previously `array[0]` produced a load
operation, which may not be eliminated and may access the array out of
range. With proper lvalue support it only computes the address and no
array access occurs.