Avoid rebuilding errorneous kernels
In some cases, such as when kernels are nested inside functions,
launches can trigger rebuilding a same kernel, which caused issues
when wanting to unit test kernel launches that were expected to
error out. The same exception would be thrown over and over again
instead of reaching the ones of the following kernels being tested.