Return default value if @wp.func might not return on all paths
We turn `if(b): ... else: ...` statements into `if(b) {...} if(!b) {...}`,
which makes it non-trivial for the C++ compiler to check if all paths of
a non-void function return a value. This change silences warnings such as
"non-void function does not return a value in all control paths".

Note that Python developers are used to getting `None` when a return is
omitted, so this closely matches the expected behavior. It also shouldn't
be left to the C++ compiler to produce `@wpfunc` compilation warnings.