Discern between value copies and assignment
Similar to C++ copy constructors and assignment operators we should
discern between `=` used for initializing a new variable or for
overwriting an existing variable. Copies are differentiable, while
assignments generally are not.

Previously `copy()` was used for both, and `adj_copy()` had a confusing
comment about not being differentiable but not actually being 'nop'.
`test_grad.test_for_loop_grad` expects differentiability despite
mutating a variable within a dynamic loop, but it's part of reversible
accumulation.