Determine if a function can be exported without raising exceptions
Previously for each builtin function we'd call the `value_func` with
empty arguments to determine if it returned a tuple. This often raised
an exception. This change passes the builtin's `value_type` to the
`Function` as well, which for all simple functions indicates whether a
tuple is returned as well.