Fix support for function calls in range()
Support for the `type()` operator in kernels introduced a regression for
`range()` expressions that contain function calls. In the case of e.g.
`wp.min()` the full function name is not an `ast.Name` but rather an
`ast.Attribute`, and in the case of just `min()` we should not
immediately treat it as an invalid operator but continue parsing it as a
potential builtin function call.