Use only the source code used for AST parsing
`adj.raw_source` was only used for annotating the generated code with
the original Python code. It is identical to `adj.source` except the
latter is dedented and not split into lines yet. The original
indentation doesn't matter because it is stripped.

Other than eliminating `adj.raw_source` the code was restructured for
readability.