Support building pip wheels for Linux AArch64
Previously we relied solely on the extension of the `warp/bin/`
libraries to determine the wheel's binary distribution platform. This no
longer works since Linux builds for x86-64 and AArch64 both use the
`.so` extension.

Instead this change relies on the libraries to be placed in a
subdirectory of `warp/bin/` which indicates one of (currently) four
platforms:
- windows-x86_64
- linux-x86_64
- linux-aarch64
- macos-universal

CD systems are responsible for putting the libraries in these
subdirectories.

Creating a wheel from a local build is still supported, by assuming the
wheel's architecture should match the system's architecture.