Remove misleading zero constructors
Constructors of the form `CUDA_CALLABLE foo_t(int) {}` existed
with the intention of allowing zero-initialization using the syntax
`foo_t foo = 0;` in the codegen of ajoint functions, however these calls
would do nothing since no member initializers were defined.

The result was only guaranteed to be initialized to zero for structs
that had their members defined with a default value such as:

```
struct foo_t
{
     int data[3] = {}; //< default value.

     CUDA_CALLABLE foo_t(int) {}
};
```

Since the `foo_t foo = 0;` doesn't seem to be in use in our adjoint
functions, we might as well get rid of it altogether.