Fix propagation of gradients on aliased copy
Copying a value to a new variable should not nullify any already
accumulated gradient.

Note this is in contrast with assigning a new value to an existing
variable (like in a loop), which isn't differentiable unless it's a
read-modify-write operation which does overwrite the gradient. Hence
the difference between `adj_copy()` and `adj_assign()`.

Issue: NVIDIA/warp#202