Fix stubs.py generation
Commit `11a4bfa008156086299ec2b33d11309c22a35924` removed defaults for
the builtins' `value_funct`. This was fine for the generation of
`exports.h` since this no longer relies on passing `None` to these
functions, but the generation of `stubs.py` (as part of `build_docs.py)
still needs it.

The removal of the duplicate of
`def select(cond: bool, arg1: Any, arg2: Any):` is a result of the
`6d47e57794242d81ed92c582b0b7e99a5386f141` commit and expected.