Upgrade building LLVM-from-source to version 18.1.3
This newer version addresses a miscompile affecting `example_cloth.py`
when running it on the CPU.

The default build still uses pre-compiled LLVM 15 static libraries
fetched from packman, for now, but when we're building LLVM from source
we can use the more recent 18.1.3 version.

Note that while this still works with GCC 7.4, the static libraries
built with that version of GCC have an ABI incompatibility with at least
GCC 9 and 10. This affects tests such as `test_copy.py`. Use of
`-fabi-version=11` does not address this issue. This isn't a problem for
from-source builds as long as these static libraries are not attempted
to be used with newer versions of GCC (i.e. not stored in packman).