Link in librt explicitly
This prevents build errors such as 'undefined reference to `shm_open`'
on older CentOS 7 images for AArch64.

Prior to glibc 2.17 the functions that are part of the POSIX Realtime
Extension resided in the separate librt library. It's a no-op to link it
for newer versions of glibc, but it doesn't exist on macOS.