Infer the number of values based on `value_type`
We want to parse arguments passed to built-in functions the same way
than Python does it, and then augment `value_func` by passing it these
arguments as-is instead of using the current parameters `arg_types`,
`kwds`, and `templates`.

To call `value_func` with the arguments given by the users, we first
need to parse/bind these arguments using the exact function's signature,
which can be known after resolving it with `Adjoint.resolve_func()`.

As a result, we need to avoid calling `value_func` before resolving
the built-in function, and since this instance was only called for
the overloads of `wp.tid()` that return mutliple values and already
provide the actual returned signature in `value_type`, then we can
easily substitute this specific call to `value_func` with the length
of `value_type`. If it's not sufficient in the future, an alternative
could be to expose a new `value_count` parameter to `add_bultin()`.