Increase PTX target architecture up to sm_75 (from sm_70)
The `sm_75` target is introduced in PTX ISA 6.3, and is supported by
CUDA 10.0 on devices with Compute Capability 7.5 (Turing architecture /
GeForce GTX 16 Series) with driver version r400 and up.

It enables the use of `sm_75` features we might make use of in the near
future.

Note that we currently require a CUDA 11.3 capable driver as a bare
minimum (see `WP_CUDA_DRIVER_VERSION`), corresponding with r465. So we
could allow up to `sm_86` at this point:
https://docs.nvidia.com/cuda/parallel-thread-execution/index.html#release-notes-ptx-release-history
For now `sm_75` seems sufficient and is more conservative to avoid any
unforeseen complications.

`ptx_target_arch` was introduced here:
https://github.com/NVIDIA/warp/commit/443aa6e300afc5032014772fac62263d5367534b