Lift the exact version req of omni.warp.core
Trying to enable `omni.warp` after launching Kit is not possible when
the exact version of `omni.warp.core` required differs from the one
being shipped with Kit.

To allow `omni.warp` to be loaded regardless of what version
of `omni.warp.core` is pinned by Kit, we remove the exact version
requirement.

This means that the versions of `omni.warp` and `omni.warp.core`
won't necessarily be in sync anymore, which could in theory lead to
compatibility issues. However, in practice, we expect any new feature
in `omni.warp` to not require the latest version of `omni.warp.core`,
and we also expect `omni.warp.core` to be stable.

Addresses GH-702.