Upgrade the LLVM source build to version 21.1.0
When building LLVM from source using the --build_llvm flag, use the
latest release version 21.1.0 instead of 18.1.3.

Note that the prebuild LLVM libraries use version 15.0.7 and 18.1.3.
Upgrading the from-source build allows to quickly check whether an issue
might be a compiler bug, and fewer changes will be required when an
upgrade becomes necessary.