Avoid static memory for tiles on AArch64 CPUs
LLVM does not fully support the large code model for JIT-compiled code
on AArch64. It uses ADRP instructions to access the GOT for static data,
but it can only be 4 GiB removed from the PC. For platforms where
allocations are highly randomized, different sections may end up further
apart than that.

To work around this, we reserve the 'shared' memory on the stack
instead. To make it accessible for allocating tiles anywhere, we store
the pointer to it in the x28 register, and instruct LLVM to reserve this
register.

x28 is the last callee-saved register in the AArch64 calling convention.
This means that if we call any external functions (i.e. not JIT-compiled
by our LLVM instance) and they need to use x28 they'll restore its value
on return.

Ensuring that our kernel restores the value of x28 upon return will be
handled in a subsequent commit.