Encapsulate shared tile storage allocation into a class
To make the allocation of shared memory for tiles more uniform between
CPU and CUDA kernels, we create a SharedTileStorage class of which an
instance is created on entry of either kernel type.

Note that initialization of the allocator now happens only once, instead
of for every task (CPU) or block (GPU). This is correct since shared
memory for tiles is relinquished when they go out of scope so the
allocator goes back to the original empty state.