Allocate shared tile memory on the stack for any CPU kernel
Instead of doing things differently between AArch64 and other CPU
architectures, always use stack memory.

This is more in line with the concept of shared memory. If Warp used
multiple threads and SIMD on the CPU, we'd need to share the memory
between lanes but not make it global to all threads.

We'd need thread_local variables to work properly in the JIT engine to
fully achieve this, or be able to reserve a register like on AArch64,
but it's a step in the right direction.