Restore the AArch64 x28 register on kernel exit
The x28 register is callee-saved according to the AArch64 ABI. This
means that upon exiting our JIT-compiled kernel we're responsible for
restoring the value it had upon entry.

Note that a caller-saved register wouldn't pose this issue but we'd have
to store and restore its value any time we called an external function,
i.e. one not part of our JIT module where we reserved x28 for our own
usage.