Eliminate static variables and the ambiguous WP_TILE_SHARED macro
While it's tempting to use a macro to write compact code that should
work on the GPU or CPU, the equivalent of `__shared__` on the CPU is
ambiguous and/or poses issues.

`static` is only correct for single-threaded execution on the CPU, and
is poorly supported by the LLVM JIT (okay on x86-64 when using large
code model, quite broken on AArch64). `thread_local` is in theory the
closest equivalent, but is explicitly not supported by the JIT.

Hence we need case-by-case workarounds (e.g. see the SharedTileStorage
implementation for AArch64) and it's better to see explicitly where
`__shared__` is being used and what is used on the CPU side instead.
Automatic storage duration is fine for variables that don't need to
persist.