Make use of stack memory for tile allocation configurable
`wp.config.enable_tiles_in_stack_memory` controls whether we use the new
approach of using stack memory that is allocated at kernel entry for
"shared" tile allocation on the CPU. When disabled we revert to using
static memory as before, which relies on the JIT compiler's support
(buggy on AArch64 but fine on x86-64).

In this commit we disable it for testing the fallback path. In the next
commit it will be re-enabled.

# Conflicts:
#	warp/native/clang/clang.cpp