Fix atomic floating-point min and max operations
The previous implementation had a bug where it would repeat the while
loop another time after already having written the new value to memory
successfully, because it compared the old memory value returned from the
CAS against the second value. The subsequent iteration would then have
the CAS return the new updated memory value (assuming no concurrent
operation happened). In other words this implementation always returned
the new minimum value, not the old memory value.

The new implementation is based more directly on
https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html?highlight=atomiccas#atomic-functions
where the loop is terminated when the value returned by the CAS matches
the previously read value, indicating the write operation successfully
atomically updated the memory (i.e. no other thread changed it).

It's possible that this "foolproof" implementation is suboptimal. In the
case of a min operation, if the value read from memory is already less
than the second value we might be able to avoid the potentially
expensive CAS operation. This was the case for the previous
implementation. However, rigorous analysis is required to validate this,
which is best done in a separate MR which can easily be reverted.