Optimize atomic floating-point min and max operations (GH-1069)
For an `atomic_min(address, val)` operation, if `val` isn't less than
the old value at `address`, we can skip trying to write the minimum of
the two into `address` since it's already the minimum.

Note that while the load operation and comparison are not atomic (i.e.
there could be a new value written to memory in between the two, which
can alter the result of the min/max, it is still equivalent to as-if the
comparison executed immediately and the whole `atomic_min` finished
before the write happened (and is thus indiscernible from fully atomic
execution, without relying on a race condition).